Move the literal to a context
